requirements.txt
flask
flask-cors
requests
discord.py